Add support for spot-fleet bootstrap
This allows using a spot fleet to create the bootstrap instance.  When instance
prices are low, this can result in savings over the on-demand cost of a
t2.small instance (amazing, to be honest), and the bootstrap server is probably
OK to handle the possible termination and re-creation that can occur with spot
instances.

A new set of variables are added to control if we should use this, and what
price to set.  We also need a variable to control the expiration time, since we
can't compute it ourselves (at least, not easily).

BUG: Setting "associate_public_ip_address", while not having any real effect in
our config (because we set that previously at the subnet level) causes
Terraform to generate a config that fails to run.  This needs to be reported.

TODO: Make multiple launch configurations for different subnet/instance-type
combinations.